fix(opencode-agent): stand down from CI-fixing a settled pull request
Recorded while fixing S2-7 and fixed here using the lookup that introduced.
`applyCiTrigger` now asks `findPullRequest` what became of the branch and stands
down for merged, closed and absent -- the third because a fix with no pull
request has nowhere to go. Standing down spends no CI-fix attempt, so a reopened
pull request still arrives with its budget intact.

The lookup sits after the "already reported" short-circuit, so a pull request
whose budget is spent and reported does not pay for an API call on every later
red run. It sits before the budget notice, so "I have stopped trying to fix CI"
is not posted to work that already landed.

Deliberately silent, unlike the spent-budget notice. That one breaks silence
because a maintainer is waiting on work that stopped; here the work has landed
or been rejected, the issue is already COMPLETE, and CI fires on every push.

Two things this pulled in. The orchestrator crossed the 300-line `max-lines`
limit, which is a design signal rather than something to compress: the trigger
layer moved to `src/triggers.ts`, leaving `orchestrator.ts` with the phase
cascade, and the shared `RunResult` moved to `types.ts` to keep the two acyclic.
And the test harness's `findPullRequest` did not know about pull requests the
harness itself had created, so COMPLETE states had no findable PR -- a fake that
disagreed with GitHub.

Refs opencode-agent/ROADMAP.md S2-11
